download document feature: route created | controller\DocumentController -> download method created | services\DocumentService -> downloadDocument method created | exception\DocumentException -> cannotView & cannotDelete methods created | model\Document -> canView & canDelete methods created

+++ app/Http/Controllers/DocumentController.php
@@ -2,11 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
+use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use App\Exceptions\DocumentException;
 use App\Services\DocumentService;
 use App\Traits\ApiResponse;
 use App\Models\Document;
+use Illuminate\Support\Facades\Storage;
 
 class DocumentController extends Controller
 {
@@ -15,10 +18,45 @@ class DocumentController extends Controller
     public function __construct(protected readonly DocumentService $documentService) {}
 
     /**
+     * Download a document.
+     */
+    public function download(
+        Request $request,
+        Document $document
+    ): BinaryFileResponse {
+        $user = $request->user();
+
+        if (!$document->canView($user)) {
+            throw DocumentException::cannotView(
+                documentId: $document->id,
+                userId: $user->id
+            );
+        }
+
+        $fileInfo = $this->documentService->downloadDocument($document);
+
+        return response()->download(
+            $fileInfo['path'],
+            $fileInfo['name']
+        );
+    }
+
+    /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Document $document): JsonResponse
-    {
+    public function destroy(
+        Request $request,
+        Document $document
+    ): JsonResponse {
+        $user = $request->user();
+
+        if (!$document->canDelete($user)) {
+            throw DocumentException::cannotDelete(
+                documentId: $document->id,
+                userId: $user->id
+            );
+        }
+
         try {
             $this->documentService->deleteDocument($document);
